Made most of analytics internal, and moved DiscoverAnalytics into analytics module
