+++ feature/discover/src/main/java/org/mozilla/social/feature/discover/DiscoverModule.kt
@@ -4,6 +4,7 @@ import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 import org.mozilla.social.common.commonModule
+import org.mozilla.social.core.analytics.DiscoverAnalytics
 import org.mozilla.social.core.analytics.analyticsModule
 import org.mozilla.social.core.navigation.navigationModule
 import org.mozilla.social.core.usecase.mozilla.mozillaUsecaseModule
@@ -18,5 +19,4 @@ val discoverModule =
         )
 
         viewModelOf(::DiscoverViewModel)
-        singleOf(::DiscoverAnalytics)
     }
