+++ feature/discover/src/main/java/org/mozilla/social/feature/discover/DiscoverViewModel.kt
@@ -7,6 +7,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import org.mozilla.social.common.Resource
+import org.mozilla.social.core.analytics.DiscoverAnalytics
 import org.mozilla.social.core.analytics.utils.ImpressionTracker
 import org.mozilla.social.core.model.Recommendation
 import org.mozilla.social.core.navigation.NavigationDestination
